% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom.R
\name{compldest}
\alias{compldest}
\title{Estimates of composite LD based either on genotype estimates or
genotype likelihoods.}
\usage{
compldest(ga, gb, K, pen = 2)
}
\arguments{
\item{ga}{One of two possible inputs: (i) A vector of counts, containing the genotypes for each individual
at the first locus; or (ii) A matrix of genotype log-likelihoods
at the second locus. The rows index the individuals and the columns
index the genotypes. That is \code{ga[i, j]} is the genotype
likelihood of individual \code{i} for genotype \code{j}.}

\item{gb}{One of two possible inputs: (i) A vector of counts, containing the genotypes for each individual
at the second locus; or (ii) A matrix of genotype log-likelihoods at
the second locus. The rows index the individuals and the columns
index the genotypes. That is \code{ga[i, j]} is the genotype
likelihood of individual \code{i} for genotype \code{j}.}

\item{K}{The ploidy of the species. Assumed the same for all individuals.}

\item{pen}{The penalty to be applied to the likelihood. You can think about
this as the prior sample size.}
}
\value{
A vector with some or all of the following elements:
\describe{
\item{\code{D}}{The MLE of D.}
\item{\code{D_se}}{The standard error of the estimate of D.}
\item{\code{r2}}{The MLE of the squared Pearson correlation.}
\item{\code{r2_se}}{The standard error of the estimate of the
squared Pearson correlation.}
\item{\code{r}}{The MLE of the Pearson correlation.}
\item{\code{r_se}}{The standard error of the estimate of the
Pearson correlation.}
\item{\code{z}}{The Fisher-z transformation of \code{r}.}
\item{\code{z_se}}{The standard error of the Fisher-z
transformation of \code{r}.}
\item{\code{p_ab}}{The estimated haplotype frequency of ab.}
\item{\code{p_Ab}}{The estimated haplotype frequency of Ab.}
\item{\code{p_aB}}{The estimated haplotype frequency of aB.}
\item{\code{p_AB}}{The estimated haplotype frequency of AB.}
}
}
\description{
This function will estimate the composite LD between two loci, either
using genotype estimates or using genotype likelihoods. The resulting
measures of LD are generalizations of Burrow's "composite" LD measure.
}
\examples{
set.seed(1)
n <- 100
K <- 6
ga <- stats::rbinom(n = n, size = K, prob = 0.5)
gb <- stats::rbinom(n = n, size = K, prob = 0.5)
ga <- t(sapply(ga, stats::dnorm, x = 0:K, sd = 1, log = TRUE))
gb <- t(sapply(gb, stats::dnorm, x = 0:K, sd = 1, log = TRUE))
pen <- 2

}
\author{
David Gerard
}
