% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLD.R
\name{dataMaster}
\alias{dataMaster}
\title{dataMaster}
\usage{
dataMaster(ldout)
}
\arguments{
\item{ldout}{Output from code{ldest::ldfast()} function.}
}
\description{
This function takes an ld output from code{ldfast} and converts the data into
a dataframe.
}
