% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{em_jointgeno}
\alias{em_jointgeno}
\title{EM algorithm to estimate joint genotype frequencies.}
\usage{
em_jointgeno(p, pgA, pgB, alpha, maxit = 500L, tol = 0.01, verbose = FALSE)
}
\arguments{
\item{p}{A matrix of proportions. The initialization of the joint genotype
frequencies. \code{p[i,j]} is the initialization of the probability
of genotype \code{i-1} on locus 1 and genotype \code{j-1} on locus 2.}

\item{pgA}{The matrix of genotype log-likelihoods for locus 1.
The rows index the individuals and the columns index the genotypes.}

\item{pgB}{The matrix of genotype log-likelihoods for locus 2.
The rows index the individuals and the columns index the genotypes.}

\item{alpha}{A matrix of prior sample sizes used as the penalty.}

\item{maxit}{The maximum number of EM iterations.}

\item{tol}{The convergence tolerance.}

\item{verbose}{Should we output the progress of each iteration (\code{TRUE})
or not (\code{FALSE})?}
}
\description{
Implements an EM algorithm to calculate the joing distribution of dosages
using genotype likelihoods.
}
\author{
David Gerard
}
