% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.R
\name{est_prior_mat}
\alias{est_prior_mat}
\title{Estimate prior genotype probabilities by maximum likelihood}
\usage{
est_prior_mat(gl, method = c("pnorm", "general"), log = TRUE)
}
\arguments{
\item{gl}{A three dimensional array of genotype \emph{log}-likelihoods.
Element \code{gl[i, j, k]} is the genotype log-likelihood of dosage
\code{k} for individual \code{j} at SNP \code{i}.}

\item{method}{Should we assume the class of general categorical prior
distributions (\code{"general"}) or the class of proportional
normal distributions (\code{"pnorm"})?}

\item{log}{A logical. Should we log the results or not?}
}
\value{
A matrix of prior genotype probabilities. Element \code{(i, k)}
is the prior probability of genotype \code{k} at locus \code{i}.
}
\description{
We maximize the marginal likelihood to obtain estimates of the prior
probabilities of each genotype at each locus. The general categorical
class uses the method of Li (2011). The proportional normal class uses
the method of Gerard and Ferrão (2020).
}
\examples{
data("glike")
est_prior_mat(glike)

}
\references{
\itemize{
\item{Gerard, D., & Ferrão, L. F. V. (2020). Priors for genotyping polyploids. Bioinformatics, 36(6), 1795-1800. \doi{10.1093/bioinformatics/btz852}}
\item{Li, H. (2011). A statistical framework for SNP calling, mutation discovery, association mapping and population genetical parameter estimation from sequencing data. \emph{Bioinformatics}, 27(21), 2987-2993. \doi{10.1093/bioinformatics/btr509}}
}
}
\seealso{
\describe{
\item{\code{\link{gl_to_gp}()}}{For converting genotype likelihoods to
genotype posteriors using prior matrices.}
}
}
\author{
David Gerard and D. Thomas Scartz
}
