% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldgeno.R
\name{format_lddf}
\alias{format_lddf}
\title{Format an element of \code{\link{mldest_geno}()} or
\code{\link{mldest_genolike}()} into an
upper-triangular matrix.}
\usage{
format_lddf(
  obj,
  element = c("r2", "r2_se", "D", "D_se", "Dprime", "Dprime_se", "r", "r_se", "p_ab",
    "p_Ab", "p_aB", "p_AB")
)
}
\arguments{
\item{obj}{An object of class \code{lddf}, usually output from
running either \code{\link{mldest_geno}()} or
\code{\link{mldest_genolike}()}.}

\item{element}{Which element in \code{obj} should we format into an
upper-triangular matrix?}
}
\description{
Formats the correlation estimates and standard errors output
from running \code{\link{mldest_geno}()} or \code{\link{mldest_genolike}()}
into a more conventional upper-triangular matrix.
}
\author{
David Gerard
}
