% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glsim.R
\name{geno_sim}
\alias{geno_sim}
\title{Simulate genotypes with a banded correlation structure.}
\usage{
geno_sim(rho, alpha, nloc, ploidy)
}
\arguments{
\item{rho}{The correlation between adjacent loci. If loci are \code{a}
positions apart, then the correlation between them is \code{rho^a}.}

\item{alpha}{The allele frequency at each locus.}

\item{nloc}{The number of loci to simulate.}

\item{ploidy}{The ploidy of the individual.}
}
\value{
A vector of length \code{nloc} with the dosage of the individual,
with a provided correlation structure.
}
\description{
Simulates genotypes at biallelic loci where each adjacent pair has
correlation \code{rho} and each locus has allele frequency \code{alpha}.
}
\examples{
set.seed(1)
hap_sim(0.5, 0.5, 50)

}
\author{
David Gerard and Hanwei Hu
}
