% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genolike_em}
\alias{genolike_em}
\title{EM algorithm to estimate haplotype frequencies}
\usage{
genolike_em(p, pgA, pgB, alpha, maxit = 100L, tol = 0.001, verbose = FALSE)
}
\arguments{
\item{p}{A vector of length 4. The intialization for the
haplotype frequencies.}

\item{pgA}{The matrix of genotype log-likelihoods for locus 1.
The rows index the individuals and the columns index the genotypes.}

\item{pgB}{The matrix of genotype log-likelihoods for locus 2.
The rows index the individuals and the columns index the genotypes.}

\item{alpha}{The prior sample size used in the penalty.}

\item{maxit}{The maximum number of EM iterations.}

\item{tol}{The convergence tolerance.}

\item{verbose}{Should we output more (\code{TRUE}) or less
(\code{FALSE}).}
}
\description{
This runs an EM algorithm to obtain the maximum likelihood estimates
of the haplotype frequencies for two loci when one has access
to genotype likelihoods.
}
\author{
David Gerard
}
