% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.R
\name{gl_to_gp}
\alias{gl_to_gp}
\title{Normalize genotype log-likelihoods to posterior probabilities.}
\usage{
gl_to_gp(gl, prior_mat = "uniform")
}
\arguments{
\item{gl}{A three dimensional array of genotype \emph{log}-likelihoods.
Element \code{gl[i, j, k]} is the genotype log-likelihood of dosage
\code{k} for individual \code{j} at SNP \code{i}.}

\item{prior_mat}{A matrix of log-prior probabilities for each genotype.
\code{prior_mat[i, k]} is the log-prior probability of genotype \code{k}
at locus \code{j}. Can also be either \code{"uniform"} for a uniform
prior for all loci, or \code{"estimate"} to adaptively estimate the
prior using \code{\link{est_prior_mat}()}. Default is a uniform
prior.}
}
\value{
A three-dimensional array, of the same dimensions as \code{gl},
containing the posterior probabilities of each dosage. This is not
on the log-scale.
}
\description{
This will take genotype log-likelihoods and a log-prior vector
and return genotype posteriors. The default corresponds to using a
naive discrete uniform prior over the genotypes.
It is not generally recommended to use a uniform prior.
}
\examples{
data("glike")
class(glike)
dim(glike)
gp1 <- gl_to_gp(glike, prior_mat = "uniform")
gp2 <- gl_to_gp(glike, prior_mat = "estimate")

}
\seealso{
\describe{
\item{\code{\link{est_prior_mat}()}}{For adaptively estimate the matrix
of prior probabilities from genotype likelihoods.}
}
}
\author{
David Gerard
}
