% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glsim.R
\name{glsim}
\alias{glsim}
\title{Simulate genotype likelihood array for individuals given a banded correlation structure.}
\usage{
glsim(
  rho,
  alpha,
  nloc,
  ploidy,
  nind,
  rdepth = 10,
  seq = 0.001,
  bias = 1,
  od = 0.001
)
}
\arguments{
\item{rho}{The correlation between adjacent loci. If loci are \code{a}
positions apart, then the correlation between them is \code{rho^a}.}

\item{alpha}{The allele frequency at each locus.}

\item{nloc}{The number of loci to simulate.}

\item{ploidy}{The ploidy of the individual.}

\item{nind}{The number of individuals to sample.}

\item{rdepth}{The read-depth of the sequencing reads used to create the
genoype likelihoods. Lower means greater uncertainty.}

\item{seq}{The sequencing error rate.}

\item{bias}{The allele bias.}

\item{od}{The overdispersion rate.}
}
\description{
Simulates genotypes using \code{\link{geno_sim}()}, then simulates
sequencing data using the model of Gerard et al (2018), then fits
\code{\link[updog]{flexdog}()} on these data to obtain genotype likelihoods.
}
\examples{
glsim(rho = 0.9, alpha = 0.5, nloc = 10, ploidy = 4, nind = 25)

}
\references{
\itemize{
\item{Gerard, D., Ferrão, L. F. V., Garcia, A. A. F., & Stephens, M. (2018). Genotyping polyploids from messy sequencing data. Genetics, 210(3), 789-807. \doi{genetics.118.301468}}
}
}
\author{
David Gerard and Hanwei Hu
}
