% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glsim.R
\name{glsim_pairwise}
\alias{glsim_pairwise}
\title{A function to simulate genotype log-likelihoods under a given correlation under HWE for two loci}
\usage{
glsim_pairwise(
  rho,
  nind,
  ploidy,
  pA,
  pB,
  rdepth = 10,
  seq = 0.001,
  bias = 1,
  od = 0.001,
  est = FALSE
)
}
\arguments{
\item{rho}{The true correlation.}

\item{nind}{The number of individuals to simulate.}

\item{ploidy}{The ploidy of the species.}

\item{pA}{The allele frequency of the first locus.}

\item{pB}{The allele frequency of the second locus.}

\item{rdepth}{The read depth to simulate (larger means more certain).}

\item{seq}{The sequencing error rate.}

\item{bias}{The allele bias.}

\item{od}{The overdispersion parameter.}

\item{est}{A logical. Should updog estimate the hyperparameters (\code{TRUE})
or use the true hyperparameter values (\code{FALSE})?}
}
\value{
A list with some or all of the following elements
\describe{
\item{lA}{The genotype log-likelihoods at the first locus. \code{lA[i, k]}
is the genotype log-likelihood for individual \code{i} at genotype \code{k - 1}.}
\item{lB}{The genotype log-likelihoods at the second locus. \code{lB[i, K]}
is the genotype log-likelihood for individual \code{i} at genotype \code{k - 1}.}
\item{rho}{The true Pearson correlation that the user provided.}
\item{delta}{The corresponding LD coefficient.}
\item{gA}{The vector of true genotypes for locus A}
\item{gB}{The vector of true genotypes for locus B}
\item{pA}{The matrix of posterior genotype probabilities for locus A}
\item{pB}{The matrix of posterior genotype probabilities for locus B}
}
}
\description{
A user inputs the correlation (\code{rho}), the number of individuals
(\code{nind}), the ploidy (\code{ploidy}), the allele frequencies at two
loci (\code{pA} and \code{pB}), and the "read-depth" (larger corresponds
to being more sure of the genotypes) (\code{rdepth}). Then this
function will  simulate some genotype log-likelihoods.
}
\examples{
glout <- glsim_pairwise(rho = 0.9, nind = 100, ploidy = 6, pA = 0.5, pB = 0.5)

cor(glout$gA, glout$gB)

}
\author{
David Gerard
}
