% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gp}
\alias{gp}
\title{Posterior probabilities from \code{\link{uit}}}
\format{
A three-dimensional \code{array} object.
}
\source{
\doi{10.1371/journal.pone.0062355}
}
\usage{
gp
}
\description{
Contains an array of posterior probabilities of the genotypes from
the \code{\link{uit}} dataset. Element \code{gp[i, j, k]} is the
posterior probability of dosage \code{k-1} for individual \code{j}
at SNP \code{i}.
}
\references{
\itemize{
\item{Uitdewilligen, Jan GAML, Anne-Marie A. Wolters, B. Bjorn, Theo JA Borm, Richard GF Visser, and Herman J. Van Eck. "A next-generation sequencing method for genotyping-by-sequencing of highly heterozygous autotetraploid potato." \emph{PloS one} 8, no. 5 (2013): e62355. \doi{10.1371/journal.pone.0062355}}
}
}
\seealso{
\code{\link{uit}} for the full \code{multidog()} fit.
}
\keyword{datasets}
