% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glsim.R
\name{hap_sim}
\alias{hap_sim}
\title{Simulate haplotypes with a banded correlation structure.}
\usage{
hap_sim(rho, alpha, nloc)
}
\arguments{
\item{rho}{The correlation between adjacent loci. If loci are \code{a}
positions apart, then the correlation between them is \code{rho^a}.}

\item{alpha}{The allele frequency at each locus.}

\item{nloc}{The number of loci to simulate.}
}
\value{
A vector of length \code{nloc} with 0's and 1's, with a
provided correlation structure.
}
\description{
Simulates haplotypes at biallelic loci where each adjacent pair has
correlation \code{rho} and each locus has allele frequency \code{alpha}.
}
\examples{
set.seed(1)
hap_sim(0.5, 0.5, 50)

}
\author{
David Gerard and Hanwei Hu
}
