% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glsim.R
\name{hap_truecor}
\alias{hap_truecor}
\title{Return the true correlation from \code{\link{hap_sim}()}.}
\usage{
hap_truecor(rho, nloc)
}
\arguments{
\item{rho}{The correlation between adjacent loci. If loci are \code{a}
positions apart, then the correlation between them is \code{rho^a}.}

\item{nloc}{The number of loci to simulate.}
}
\value{
A matrix of size \code{nloc} by \code{nloc} that contains the
true correlation of the data simulated in \code{\link{hap_sim}()}.
}
\description{
The correlation is \code{rho^a}, where \code{a} is the number of
positions away two SNPs are.
}
\examples{
hap_truecor(0.9, 10)

}
\author{
David Gerard and Hanwei Hu
}
