% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.R
\name{hsm}
\alias{hsm}
\title{Half-sample mode estimator}
\usage{
hsm(x, k = 0.5)
}
\arguments{
\item{x}{A vector of values}

\item{k}{Fraction to discard each iteration.}
}
\value{
The mode estimate.
}
\description{
Implementation of the half-sample mode estimator of
Robertson & Cryer (1974).
}
\details{
Notes that in the case of ties, this version automatically chooses the
smallest of the ties. So, don't use this with ties.
}
\examples{
set.seed(1)
x <- stats::rnorm(100)
hsm(x)

}
\references{
\itemize{
\item{Robertson, T., & Cryer, J. D. (1974). An iterative procedure for estimating the mode. Journal of the American Statistical Association, 69(348), 1012-1016. \doi{10.1080/01621459.1974.10480246}}
}
}
\author{
David Gerard
}
