% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{is.lddf}
\alias{is.lddf}
\title{Tests if an argument is a \code{lddf} object.}
\usage{
is.lddf(x)
}
\arguments{
\item{x}{Anything.}
}
\value{
A logical. \code{TRUE} if \code{x} is a \code{lddf} object,
and \code{FALSE} otherwise.
}
\description{
Tests if an argument is a \code{lddf} object.
}
\examples{
is.lddf("anything")
# FALSE

}
\author{
David Gerard
}
