% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldgeno.R
\name{ldest_geno}
\alias{ldest_geno}
\title{Estimate LD directly from the genotypes}
\usage{
ldest_geno(ga, gb, K)
}
\arguments{
\item{ga}{A vector of counts, containing the genotypes for each individual
at the first locus.}

\item{K}{the ploidy of the species.}

\item{A}{vector of counts, containing the genotypes for each individual
at the second locus.}
}
\description{
Estimate LD directly from the genotypes
}
\author{
David Gerard
}
