% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldgeno.R
\name{ldest_genolike}
\alias{ldest_genolike}
\title{Estimate LD from genotype likelihoods.}
\usage{
ldest_genolike(genlike1, genlike2)
}
\arguments{
\item{genlike1}{A matrix of genotype log-likelihoods for locus 1.
The rows index the individuals and the columns index the genotypes.
The ploidy of the individuals is assumed to be the number of columns
minus 1.}

\item{genlike2}{A matrix of genotype log-likelihoods for locus 1.
The rows index the individuals and the columns index the genotypes.
The ploidy of the individuals is assumed to be the number of columns
minus 1.}
}
\description{
Estimate LD from genotype likelihoods.
}
\author{
David Gerard
}
