% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.R
\name{ldfast_naive}
\alias{ldfast_naive}
\title{Naive approach to estimate LD using posterior moments.}
\usage{
ldfast_naive(gp, type = c("r", "r2", "z", "D", "Dprime"))
}
\arguments{
\item{gp}{A three-way array with dimensions SNPs by individuals by dosage.
That is, \code{gp[i, j, k]} is the posterior probability of
dosage \code{k-1} for individual \code{j} at SNP \code{i}.}

\item{type}{What LD measure should we estimate?
\describe{
\item{\code{"r"}}{The Pearson correlation.}
\item{\code{"r2"}}{The squared Pearson correlation.}
\item{\code{"z"}}{The Fisher-z transformed Pearson correlation.}
\item{\code{"D"}}{The LD coefficient.}
\item{\code{"Dprime"}}{The standardized LD coefficient.}
}
Note that these are all \emph{composite} measures of LD (see
the description in \code{\link{ldest}()}).}
}
\value{
A list with the following elements.
\describe{
\item{\code{ldmat}}{The bias-corrected LD matrix.}
}
}
\description{
Here, we just use posterior means as a proxy for the genotypes, and calculate
these LD estimates as if the genotypes were known. This is not a good way
to do things if there is any genotype uncertainty (which is almost always
the case in polyploids).
}
\examples{
data("glike")
gp <- gl_to_gp(gl = glike, prior_mat = "estimate_pnorm")
ldfast_naive(gp = gp)

}
\author{
David Gerard
}
