% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.R
\name{ldfast_unif}
\alias{ldfast_unif}
\title{Scalable LD Calculations Assuming a Uniform Prior}
\usage{
ldfast_unif(
  gp,
  type = c("r", "r2", "z", "D", "Dprime"),
  shrinkrr = TRUE,
  se = TRUE,
  thresh = TRUE,
  upper = 10,
  mode = c("zero", "estimate")
)
}
\arguments{
\item{gp}{A three-way array with dimensions SNPs by individuals by dosage.
That is, \code{gp[i, j, k]} is the posterior probability of
dosage \code{k-1} for individual \code{j} at SNP \code{i}.}

\item{type}{What LD measure should we estimate?
\describe{
\item{\code{"r"}}{The Pearson correlation.}
\item{\code{"r2"}}{The squared Pearson correlation.}
\item{\code{"z"}}{The Fisher-z transformed Pearson correlation.}
\item{\code{"D"}}{The LD coefficient.}
\item{\code{"Dprime"}}{The standardized LD coefficient.}
}
Note that these are all \emph{composite} measures of LD (see
the description in \code{\link{ldest}()}).}

\item{shrinkrr}{A logical. Should we use adaptive shrinkage
(Stephens, 2016) to shrink the reliability ratios (\code{TRUE})
or keep the raw reliability ratios (\code{FALSE}). Defaults
to \code{TRUE}.}

\item{se}{Should we also return a matrix of standard errors (\code{TRUE})
or not (\code{FALSE})? It is faster to not return standard errors.
Defaults to \code{TRUE}.}

\item{thresh}{A logical. Should we apply an upper bound on the reliability
ratios (\code{TRUE}) or not (\code{FALSE}).}

\item{upper}{The upper bound on the reliability ratios if
\code{thresh = TRUE}. The default is a generous 10.}

\item{mode}{A character. Only applies if \code{shrinkrr = TRUE}. When using
hierarchical shrinkage on the log of the reliability ratios, should
we use zero as the mode (\code{mode = "zero"}) or estimate it using
the procedure of Robertson and Cryer (1974)
(\code{mode = "estimate"})?}
}
\value{
A list with some or all of the following elements:
\describe{
\item{\code{ldmat}}{The bias-corrected LD matrix.}
\item{\code{rr}}{The estimated reliability ratio for each SNP. This
is the multiplicative factor applied to the naive LD estimate
for each SNP.}
}
}
\description{
Here, we assume that \code{gp} was calculated assuming a discrete uniform
prior over the genotypes. This is different from \code{\link{ldfast}()}
where we assumed an adaptive prior had been used.
}
\examples{
data("glike")
gp <- gl_to_gp(gl = glike)
ldfast_unif(gp = gp, se = FALSE)
}
\seealso{
\describe{
\item{\code{\link{ldfast}()}}{For an approach when the genotype posteriors
used adaptive priors.}
\item{\code{\link{gl_to_gp}()}}{Convert genotype likelihoods to
genotype posteriors, optionally using uniform priors.}
}
}
\author{
David Gerard and Hanwei Hu
}
