% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom.R
\name{ldsimp}
\alias{ldsimp}
\title{Simple LD Estimators}
\usage{
ldsimp(ga, gb, K)
}
\arguments{
\item{ga}{Either the genotype at locus 1 or the posterior mean genotype at
locus 1.}

\item{gb}{Either the genotype at locus 2 or the posterior mean genotype at
locus 2.}

\item{K}{The ploidy of hte species. Assumed to be the same for all
individuals.}
}
\value{
A vector with some or all of the following elements:
\describe{
\item{\code{D}}{The MLE of D.}
\item{\code{D_se}}{The standard error of the estimate of D.}
\item{\code{r2}}{The MLE of the squared Pearson correlation.}
\item{\code{r2_se}}{The standard error of the estimate of the
squared Pearson correlation.}
\item{\code{r}}{The MLE of the Pearson correlation.}
\item{\code{r_se}}{The standard error of the estimate of the
Pearson correlation.}
\item{\code{z}}{The Fisher-z transformation of \code{r}.}
\item{\code{z_se}}{The standard error of the Fisher-z
transformation of \code{r}.}
\item{\code{p_ab}}{The estimated haplotype frequency of ab.}
\item{\code{p_Ab}}{The estimated haplotype frequency of Ab.}
\item{\code{p_aB}}{The estimated haplotype frequency of aB.}
\item{\code{p_AB}}{The estimated haplotype frequency of AB.}
}
}
\description{
Provides quick and simple estimates of LD and their corresponding
standard errors. These estimates are just based on the sample
moments of the genotypes or the posterior mean genotypes.
}
\details{
For large sequencing depth and large n, these estimates perform as well
as the MLE's but are much faster to calculate. For small n and, in
particular, small sequencing depth, these estimates tend to be very
biased.
}
\author{
David Gerard
}
