% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldgeno.R
\name{mldest_genolike}
\alias{mldest_genolike}
\title{Estimate all pair-wise LD's in a collection of SNPs using the genotype
likleihoods.}
\usage{
mldest_genolike(genoarray, nc = 1, reltol = 10^-8)
}
\arguments{
\item{genoarray}{An three-way array of genotype \emph{log}-likelihoods.
The first dimension indexes the SNPs, the second dimension indexes
the individuals, and the third dimension indexes the genotypes.
That is, \code{genolike_array[i, j, k]} is the genotype log-likelihood
at SNP \code{i} for individual \code{j} and dosage \code{k - 1}.
The ploidy (assumed to be the same for all individuals) is assumed to
be one minus the size of the third dimension.}

\item{nc}{The number of computing cores to use. This should never be
more than the number of cores available in your computing environment.
You can determine the maximum number of available cores by running
\code{parallel::detectCores()} in R. This is probably fine for a
personal computer, but some environments are only
able to use fewer. Ask your admins if you are unsure.}

\item{reltol}{The relative tolerance for the stopping criterion.}
}
\description{
This function will run \code{\link{ldest}()} iteratively over
all possible pairs of SNPs provided. Support is provided for parallelization
through the doParallel and foreach packages.
}
\examples{
## Simulate some data with true correlation of 0
nloci <- 5
nind  <- 100
K <- 6
genovec <- sample(0:K, nind * nloci, TRUE)
genomat <- matrix(genovec, nrow = nloci)
genoarray <- array(sapply(genovec,
                          stats::dnorm,
                          x = 0:K,
                          sd = 1,
                          log = TRUE),
                   dim = c(K + 1, nloci, nind))
genoarray <- aperm(genoarray, c(2, 3, 1)) ## loci, individuals, dosages

## Verify simulation
locnum <- sample(seq_len(nloci), 1)
indnum <- sample(seq_len(nind), 1)
genoarray[locnum, indnum, ]
stats::dnorm(x = 0:K, mean = genomat[locnum, indnum], sd = 1, log = TRUE)

## Find pairwise LD between all loci
nc <- 1
rdf <- mldest_genolike(genoarray = genoarray, nc = nc)
rdf

}
\author{
David Gerard
}
