% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldgeno.R
\name{multi_ldest_geno}
\alias{multi_ldest_geno}
\title{Estimate all pair-wise LD's in a collection of SNPs.}
\usage{
multi_ldest_geno(genomat, K, nc = 1, reltol = 10^-8)
}
\arguments{
\item{genomat}{A matrix of genotypes (allele dosages). The rows index the
SNPs and the columns index the individuals. That is, \code{genomat[i, j]}
is the allele dosage for individual \code{j} in SNP \code{i}.}

\item{K}{The ploidy of the species. Assumed to be the same for all
individuals at all SNPs}

\item{nc}{he number of computing cores to use. This should never be
more than the number of cores available in your computing environment.
You can determine the maximum number of available cores by running
\code{parallel::detectCores()} in R. This is probably fine for a
personal computer, but some environments are only
able to use fewer. Ask your admins if you are unsure.}

\item{reltol}{The relative tolerance for the stopping criterion.}
}
\description{
This function will run \code{\link{ldest_geno}()} iteratively over
all possible pairs of SNPs provided. Support is provided for parallelization
through the doParallel and foreach packages.
}
\examples{
nloci <- 5
nind  <- 100
K <- 6
nc <- 2
genomat <- matrix(sample(0:K, nind * nloci, TRUE), nrow = nloci)
rdf <- multi_ldest_geno(genomat = genomat, K = K, nc = nc)


}
\author{
David Gerard
}
