% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pbnorm_dist}
\alias{pbnorm_dist}
\title{Returns distribution of proportional bivariate normal.}
\usage{
pbnorm_dist(mu, sigma, K, log = FALSE)
}
\arguments{
\item{mu}{A vector of length 2 containing the mean.}

\item{sigma}{A 2-by-2 positive definite covariance matrix}

\item{K}{The ploidy of the individual.}

\item{log}{A logical. If \code{TRUE}, log probabilities are returned.}
}
\value{
A matrix. Element (i,j) is the (log) probability of genotype
i-1 at locus 1 and j-1 at locus 2.
}
\description{
Returns distribution of proportional bivariate normal.
}
\author{
David Gerard
}
