% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.R
\name{plot.ldfast}
\alias{plot.ldfast}
\title{Plot ldfast output method}
\usage{
\method{plot}{ldfast}(x, groupPoints = 20, ...)
}
\arguments{
\item{x}{ldfast output object.}

\item{groupPoints}{Number of groups to use.}

\item{...}{Further arguments.}
}
\description{
Plot ldfast output method
}
