% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLD.R
\name{plotLD}
\alias{plotLD}
\title{plotLD}
\usage{
plotLD(ldout, groupPoints = 20)
}
\arguments{
\item{ldout}{Output from code{ldest::ldfast()} function.}

\item{groupPoints}{the number of discrete distance groupings.}
}
\description{
This function takes an ldout from \code{ldfast} and a vector of genetic distances
and plots LD at various distances.
}
