% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.R
\name{shrink_ldfast}
\alias{shrink_ldfast}
\title{Shrink LD Estimates based on LD decay}
\usage{
shrink_ldfast(ldout, dof = 8, backTo = c("r2"))
}
\arguments{
\item{ldout}{ldfast output object.}

\item{dof}{Degrees of freedom for fitting spline.}

\item{backTo}{value to convert adjusted z to.}
}
\description{
Shrink LD Estimates based on LD decay
}
