% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLD.R
\name{sumDf}
\alias{sumDf}
\title{sumDF}
\usage{
sumDf(DFncomb, ngroup = 4)
}
\arguments{
\item{DFncomb}{Datamaster output.}

\item{ngroup}{Number of groups.}
}
\description{
This function groups the data according to genetic distance into n groups.
}
