% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnorm.R
\name{tnorm_est}
\alias{tnorm_est}
\title{Estimate parameters of truncated normal}
\usage{
tnorm_est(xbar, s, ploidy)
}
\arguments{
\item{xbar}{The sample mean.}

\item{s}{The sample standard deviation.}

\item{ploidy}{The upper bound of the truncated normal.}
}
\value{
A vector of length 2. The first element is the estimate of the
mean parameter, the second element is the estimate of the standard
deviation parameter.
}
\description{
Scalable approximation to estimate the parameters of the truncated normal
by method-of-moments / maximum likelihood (equivalent methods in this case).
We do so by the look-up table method of Cohen (1957). The lower bound of
the truncated normal is assumed to be 0. Just pre and post shift the data
if not.
}
\examples{
xbar <- 2
s <- 1
ploidy <- 6
tnorm_est(xbar = xbar, s = s, ploidy = ploidy)

}
\references{
\itemize{
\item{Cohen, A. C. (1957). On the solution of estimating equations for truncated and censored samples from normal populations. \emph{Biometrika}, 44(1/2), 225-236. \doi{10.2307/2333256}}
}
}
\author{
David Gerard
}
