% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uit}
\alias{uit}
\title{Updog fits on the data from Uitdewilligen et. al. (2013)}
\format{
An object of class \code{\link[updog]{multidog}()}.
See the documentation from the updog R package.
}
\source{
\doi{10.1371/journal.pone.0062355}
}
\usage{
uit
}
\description{
10 SNPs from the "PGSC0003DMB000000062" super scaffold were genotyped
using the \code{\link[updog]{multidog}()} function from the updog R package.
These data are the resulting output.
}
\references{
\itemize{
\item{Uitdewilligen, Jan GAML, Anne-Marie A. Wolters, B. Bjorn, Theo JA Borm, Richard GF Visser, and Herman J. Van Eck. "A next-generation sequencing method for genotyping-by-sequencing of highly heterozygous autotetraploid potato." \emph{PloS one} 8, no. 5 (2013): e62355. \doi{10.1371/journal.pone.0062355}}
}
}
\keyword{datasets}
