% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genolike_em}
\alias{genolike_em}
\title{EM algorithm to estimate haplotype frequencies}
\usage{
genolike_em(
  p,
  pgA,
  pgB,
  alpha,
  maxit = 500L,
  tol = 1e-04,
  verbose = FALSE,
  square = FALSE
)
}
\arguments{
\item{p}{A vector of length 4. The intialization for the
haplotype frequencies.}

\item{pgA}{The matrix of genotype log-likelihoods for locus 1.
The rows index the individuals and the columns index the genotypes.}

\item{pgB}{The matrix of genotype log-likelihoods for locus 2.
The rows index the individuals and the columns index the genotypes.}

\item{alpha}{The prior sample size used in the penalty.}

\item{maxit}{The maximum number of EM iterations.}

\item{tol}{The convergence tolerance.}

\item{verbose}{Should we output more (\code{TRUE}) or less
(\code{FALSE})?}

\item{square}{Should we implement squared acceleratred EM (\code{TRUE})
or not (\code{FALSE})?}
}
\description{
This runs an EM algorithm to obtain the maximum likelihood estimates
of the haplotype frequencies for two loci when one has access
to genotype likelihoods.
}
\details{
EM Squaring is performed via the algorithm of Ravi and Roland (2008)
with each iteration's projection onto the simplex using the algorithm
of Chen and Ye (2011). Though squaring is turned off be default because
it usually doesn't help.
}
\references{
\itemize{
\item{Varadhan, Ravi, and Christophe Roland. "Simple and globally convergent methods for accelerating the convergence of any EM algorithm." Scandinavian Journal of Statistics 35.2 (2008): 335-353.}
\item{Chen, Yunmei, and Xiaojing Ye. "Projection onto a simplex." arXiv preprint arXiv:1101.6081 (2011).}
}
}
\author{
David Gerard
}
