% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.R
\name{ldfast}
\alias{ldfast}
\title{Fast bias-correction for LD Estimation}
\usage{
ldfast(gp, type = c("r", "r2", "z", "D", "Dprime"))
}
\arguments{
\item{gp}{A three-way array with dimensions SNPs by individuals by dosage.
That is, \code{gp[i, j, k]} is the posterior probability of
dosage \code{k-1} for individual \code{j} at SNP \code{i}.}

\item{type}{What LD measure should we estimate?
\describe{
\item{\code{r}}{The Pearson correlation.}
\item{\code{r2}}{The squared Pearson correlation.}
\item{\code{z}}{The Fisher-z transformed Pearson correlation.}
\item{\code{D}}{The LD coefficient.}
\item{\code{Dprime}}{The standardized LD coefficient.}
}
Note that these are all \emph{composite} measures of LD.}
}
\value{
A list with the following elements:
\describe{
\item{\code{ldmat}}{The bias-corrected LD matrix.}
\item{\code{semat}}{A matrix 0f standard errors of the corresponding
estimators of LD.}
}
}
\description{
Estimates the reliability ratios from posterior marginal moments and uses
these to correct the biases in linkage disequilibrium estimation
caused by genotype uncertainty.
}
\section{Details}{


Sometimes, the reliability ratio results in a correlation greater than 1
or less than -1 (due to sample variability). We truncate these occasions
at 1 and -1, respectively.
}

\section{Mathematical formulation}{

Let
\itemize{
\item{\eqn{r} be the sample correlation of posterior mean genotypes
between loci 1 and 2,}
\item{\eqn{a1} be the sample variance of posterior means at locus 1,}
\item{\eqn{a2} be the sample variance of posterior means at locus 2,}
\item{\eqn{b1} be the sample mean of posterior variances at locus 1, and}
\item{\eqn{b2} be the sample mean of posterior variances at locus 2.}
}
Then the estimated Pearson correlation between the genotypes at
loci 1 and 2 is
\deqn{[(a1 + b1)/a1][(a2 + b2)/a2]r.}
This is the estimated LD when \code{pv = NULL}.
}

\examples{
## Load the data -----
data("gp") # posterior probs
ldout <- ldfast(gp, "r")
ldout$ldmat
ldout$semat

ldout <- ldfast(gp, "D")
ldout$ldmat
ldout$semat

ldout <- ldfast(gp, "Dprime")
ldout$ldmat
ldout$semat

}
\seealso{
\describe{
\item{\code{\link{gl_to_gp}()}}{Normalize genotype likelihoods to
posterior probabilities using naive uniform prior.}
\item{\code{\link{pvcalc}()}}{Calculate prior variances given genotype
priors.}
}
}
\author{
David Gerard
}
