% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.R
\name{ldfast}
\alias{ldfast}
\title{Fast bias-correction for LD Estimation}
\usage{
ldfast(gp, type = c("r", "r2", "z", "D", "Dprime"), se = TRUE)
}
\arguments{
\item{gp}{A three-way array with dimensions SNPs by individuals by dosage.
That is, \code{gp[i, j, k]} is the posterior probability of
dosage \code{k-1} for individual \code{j} at SNP \code{i}.}

\item{type}{What LD measure should we estimate?
\describe{
\item{\code{r}}{The Pearson correlation.}
\item{\code{r2}}{The squared Pearson correlation.}
\item{\code{z}}{The Fisher-z transformed Pearson correlation.}
\item{\code{D}}{The LD coefficient.}
\item{\code{Dprime}}{The standardized LD coefficient.}
}
Note that these are all \emph{composite} measures of LD.}

\item{se}{Should we also return a matrix of standard errors (\code{TRUE})
or not (\code{FALSE})? It is faster to not return standard errors.
Defaults to \code{TRUE}.}
}
\value{
A list with some or all of the following elements:
\describe{
\item{\code{ldmat}}{The bias-corrected LD matrix.}
\item{\code{semat}}{A matrix 0f standard errors of the corresponding
estimators of LD.}
}
}
\description{
Estimates the reliability ratios from posterior marginal moments and uses
these to correct the biases in linkage disequilibrium estimation
caused by genotype uncertainty.
}
\section{Details}{


Returns consistent and bias-corrected estimates of Linkage Disequilibrium.
The usual measures of LD are implemented: D, D', r, r2, and z
(Fisher-z of r). These are all \emph{composite} measures of LD, not
gametic measures of LD. They are always appropriate measures of association
between loci, but only correspond to gametic measures of LD when
Hardy-Weinberg equilibrium is fulfilled in polyploids.

Due to sampling varibility, the estimates sometimes lie outside of the
theoretical boundary of the parameters being estimated. In such cases,
we truncate the estimates at the boundary and return \code{NA} for the
standard errors.
}

\section{Mathematical formulation}{

Let
\itemize{
\item{\eqn{r} be the sample correlation of posterior mean genotypes
between loci 1 and 2,}
\item{\eqn{a1} be the sample variance of posterior means at locus 1,}
\item{\eqn{a2} be the sample variance of posterior means at locus 2,}
\item{\eqn{b1} be the sample mean of posterior variances at locus 1, and}
\item{\eqn{b2} be the sample mean of posterior variances at locus 2.}
}
Then the estimated Pearson correlation between the genotypes at
loci 1 and 2 is
\deqn{[(a1 + b1)/a1][(a2 + b2)/a2]r.}
This is the estimated LD when \code{pv = NULL}.
}

\examples{
## Load the data -----
data("gp") # posterior probs
ldout <- ldfast(gp, "r")
ldout$ldmat
ldout$semat

ldout <- ldfast(gp, "D")
ldout$ldmat
ldout$semat

ldout <- ldfast(gp, "Dprime")
ldout$ldmat
ldout$semat

}
\seealso{
\describe{
\item{\code{\link{gl_to_gp}()}}{Normalize genotype likelihoods to
posterior probabilities using naive uniform prior.}
}
}
\author{
David Gerard
}
