% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.R
\name{ldfast}
\alias{ldfast}
\title{Fast bias-correction for LD Estimation}
\usage{
ldfast(gp, type = c("r", "r2", "D"), pv = NULL)
}
\arguments{
\item{gp}{A three-way array with dimensions SNPs by individuals by dosage.
That is, \code{gp[i, j, k]} is the posterior probability of
dosage \code{k-1} for individual \code{j} at SNP \code{i}.}

\item{type}{What LD measure should we estimate? The Pearson correlation
(\code{type = "r"}), the squared Pearson correlation
(\code{type = "r2"}), or the LD coefficient (\code{type = "D"})?
These are all \emph{composite} measures of LD.}

\item{pv}{A vector of numerics. Optionally provided prior variances.
This should only be provided if these prior variances were estimated
adaptively.}
}
\value{
A list with the following elements:
\describe{
\item{\code{cor}}{The bias-corrected LD matrix.}
\item{\code{rr}}{The estimated reliability ratio for the
\emph{correlation}. This is how much you multiply the
Pearson correlation by to obtain the bias-adjusted Pearson
correlation estimate. Square this term to get the reliability
ratio for the LD coefficient.}
}
}
\description{
Estimates the reliability ratios from posterior marginal moments and uses
these to correct the biases in linkage disequilibrium estimation
caused by genotype uncertainty.
}
\section{Details}{


Sometimes, the reliability ratio results in a correlation greater than 1
or less than -1 (due to sample variability). We truncate these occasions
at 1 and -1, respectively.

This won't work for missing data in \code{gp} yet, but the plan is to
allow missing data in the future.
}

\section{Mathematical formulation}{

Let
\itemize{
\item{\eqn{r} be the sample correlation of posterior mean genotypes
between loci 1 and 2,}
\item{\eqn{a1} be the sample variance of posterior means at locus 1,}
\item{\eqn{a2} be the sample variance of posterior means at locus 2,}
\item{\eqn{b1} be the sample mean of posterior variances at locus 1, and}
\item{\eqn{b2} be the sample mean of posterior variances at locus 2.}
}
Then the estimated Pearson correlation between the genotypes at
loci 1 and 2 is
\deqn{[(a1 + b1)/a1][(a2 + b2)/a2]r.}
This is the estimated LD when \code{pv = NULL}.

Alternatively, suppose that we know the prior variances and we let
\itemize{
\item{\eqn{s1} be the prior variance at locus 1,}
\item{\eqn{s2} be the prior variance at locus 2, and}
\item{\eqn{t} be the sample covariance of posterior mean genotypes
between loci 1 and 2.}
}
Then the estimated covariance between genotypes at loci 1 and 2 is
\deqn{[s1/(s1 - b1)][s2/(s2-b2)]t.}
We obtain the estimate Pearson correlation by normalizing by the
prior standard deviations
\deqn{[\sqrt{s1}/(s1 - b1)][\sqrt{s2}/(s2-b2)]t.}
This is the estimated LD when \code{pv} is not \code{NULL}.
}

\examples{
## Load the data -----
data("uit") # mupdog() fits
pv <- pvcalc(uit$snpdf[, paste0("Pr_", 0:4)]) # prior probs
data("gp") # posterior probs

## Estimate squared correlation using just posterior moments -----
cout1 <- ldfast(gp = gp, type = "r2")
corrplot::corrplot(corr = cout1$cor,
                   diag = FALSE,
                   type = "upper",
                   method = "color")
hist(cout1$rr, xlab = "Reliability Ratio")

## Estimate squared correlation also using prior variances -----
cout2 <- ldfast(gp = gp, type = "r2", pv = pv)
corrplot::corrplot(corr = cout2$cor,
                   diag = FALSE,
                   type = "upper",
                   method = "color")
hist(cout2$rr, xlab = "Reliability Ratio")

## The adjustments are really close in these data -----
plot(x = cout1$rr,
     y = cout2$rr,
     xlab = "Without Prior Variance",
     ylab = "With Prior Variance")
abline(0, 1)

}
\seealso{
\describe{
\item{\code{\link{gl_to_gp}()}}{Normalize genotype likelihoods to
posterior probabilities using naive uniform prior.}
\item{\code{\link{pvcalc}()}}{Calculate prior variances given genotype
priors.}
}
}
\author{
David Gerard
}
