% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.R
\name{ldfast}
\alias{ldfast}
\title{Fast bias-correction for LD Estimation}
\usage{
ldfast(gp, type = c("r", "r2", "z", "D", "Dprime"), shrinkrr = TRUE, se = TRUE)
}
\arguments{
\item{gp}{A three-way array with dimensions SNPs by individuals by dosage.
That is, \code{gp[i, j, k]} is the posterior probability of
dosage \code{k-1} for individual \code{j} at SNP \code{i}.}

\item{type}{What LD measure should we estimate?
\describe{
\item{\code{r}}{The Pearson correlation.}
\item{\code{r2}}{The squared Pearson correlation.}
\item{\code{z}}{The Fisher-z transformed Pearson correlation.}
\item{\code{D}}{The LD coefficient.}
\item{\code{Dprime}}{The standardized LD coefficient.}
}
Note that these are all \emph{composite} measures of LD.}

\item{shrinkrr}{A logical. Should we use adaptive shrinkage to shrink
the reliability ratios (\code{TRUE}) or keep the raw reliability
ratios (\code{FALSE}). Defaults to \code{TRUE}.}

\item{se}{Should we also return a matrix of standard errors (\code{TRUE})
or not (\code{FALSE})? It is faster to not return standard errors.
Defaults to \code{TRUE}.}
}
\value{
A list with some or all of the following elements:
\describe{
\item{\code{ldmat}}{The bias-corrected LD matrix.}
\item{\code{rr}}{The estimated reliability ratio for each SNP. This
is the multiplicative factor applied to the naive LD estimate
for each SNP.}
\item{\code{semat}}{A matrix 0f standard errors of the corresponding
estimators of LD.}
}
}
\description{
Estimates the reliability ratios from posterior marginal moments and uses
these to correct the biases in linkage disequilibrium estimation
caused by genotype uncertainty.
}
\section{Details}{


Returns consistent and bias-corrected estimates of Linkage Disequilibrium.
The usual measures of LD are implemented: D, D', r, r2, and z
(Fisher-z of r). These are all \emph{composite} measures of LD, not
gametic measures of LD. They are always appropriate measures of association
between loci, but only correspond to gametic measures of LD when
Hardy-Weinberg equilibrium is fulfilled in polyploids.

Calculating standard errors and performing hierarchical shrinkage of the
reliability ratios are both rather slow operations compared to just
raw method-of-moments based estimation for LD. If you don't need
standard errors, you can double your speed by setting
\code{se = FALSE}. It is not recommended that you disable the
hierarchical shrinkage.

Due to sampling varibility, the estimates sometimes lie outside of the
theoretical boundary of the parameters being estimated. In such cases,
we truncate the estimates at the boundary and return \code{NA} for the
standard errors.
}

\section{Mathematical formulation}{

Let
\itemize{
\item{\eqn{r} be the sample correlation of posterior mean genotypes
between loci 1 and 2,}
\item{\eqn{a1} be the sample variance of posterior means at locus 1,}
\item{\eqn{a2} be the sample variance of posterior means at locus 2,}
\item{\eqn{b1} be the sample mean of posterior variances at locus 1, and}
\item{\eqn{b2} be the sample mean of posterior variances at locus 2.}
}
Then the estimated Pearson correlation between the genotypes at
loci 1 and 2 is
\deqn{[(a1 + b1)/a1][(a2 + b2)/a2]r.}
This is the estimated LD when \code{pv = NULL}.
}

\examples{
## Load the data -----
data("gp") # posterior probs
ldout <- ldfast(gp, "r", shrinkrr = FALSE)
ldout$ldmat
ldout$rr
ldout$semat

ldout <- ldfast(gp, "D")
ldout$ldmat
ldout$rr
ldout$semat

ldout <- ldfast(gp, "Dprime")
ldout$ldmat
ldout$rr
ldout$semat

}
\seealso{
\describe{
\item{\code{\link{gl_to_gp}()}}{Normalize genotype likelihoods to
posterior probabilities using naive uniform prior.}
\item{\code{\link[ashr]{ash}()}}{Function used to perform hierarchical
shrinkage on the log of the reliability ratios.}
}
}
\author{
David Gerard
}
