% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsep-package.R
\docType{package}
\name{ldsep-package}
\alias{ldsep-package}
\alias{ldsep}
\title{Linkage Disequilibrium Shrinkage Estimation for Polyploids}
\description{
Estimates pairwise linkage disequilibrium (LD), using either
genotype estimates or genotype likelihoods. Functions are then provided
for shrinkage estimation using multidimensional scaling and adaptive
shrinkage. The idea is to expand this to allow for a sliding window
estimation procedure for genome-wise LD estimation.
}
\author{
David Gerard
}
