% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsep-package.R
\docType{package}
\name{ldsep-package}
\alias{ldsep-package}
\alias{ldsep}
\title{Linkage Disequilibrium Shrinkage Estimation for Polyploids}
\description{
Estimate haplotypic or composite pairwise linkage disequilibrium
(LD), using either genotypes or genotype likelihoods. Support is
provided to estimate the popular measures of LD: the LD coefficient D,
the standardized LD coefficient D', and the Pearson correlation
coefficient r. All estimates are returned with corresponding
standard errors. These estimates and standard errors can then be used
for shrinkage estimation. The main functions are \code{\link{ldest}()},
\code{\link{mldest}()}, \code{\link{plot.lddf}()},
\code{\link{format_lddf}()}, and \code{\link{ldshrink}()}.
}
\author{
David Gerard
}
