% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsep-package.R
\docType{package}
\name{ldsep-package}
\alias{ldsep}
\alias{ldsep-package}
\title{ldsep: Linkage Disequilibrium Shrinkage Estimation for Polyploids}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Estimate haplotypic or composite pairwise linkage disequilibrium (LD) in polyploids, using either genotypes or genotype likelihoods. Support is provided to estimate the popular measures of LD: the LD coefficient D, the standardized LD coefficient D', and the Pearson correlation coefficient r. All estimates are returned with corresponding standard errors. These estimates and standard errors can then be used for shrinkage estimation. The main functions are ldfast(), ldest(), mldest(), sldest(), plot.lddf(), format_lddf(), and ldshrink(). Details of the methods are available in Gerard (2021a) \doi{10.1111/1755-0998.13349} and Gerard (2021b) \doi{10.1038/s41437-021-00462-5}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dcgerard.github.io/ldsep/}
  \item Report bugs at \url{https://github.com/dcgerard/ldsep/issues}
}

}
\author{
\strong{Maintainer}: David Gerard \email{gerard.1787@gmail.com} (\href{https://orcid.org/0000-0001-9450-5023}{ORCID})

}
\keyword{internal}
