% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldgeno.R
\name{mldest_geno}
\alias{mldest_geno}
\title{Estimate all pair-wise LD's in a collection of SNPs using the genotypes.}
\usage{
mldest_geno(genomat, K, nc = 1, reltol = 10^-8, pen = 2)
}
\arguments{
\item{genomat}{A matrix of genotypes (allele dosages). The rows index the
SNPs and the columns index the individuals. That is, \code{genomat[i, j]}
is the allele dosage for individual \code{j} in SNP \code{i}.}

\item{K}{The ploidy of the species. Assumed to be the same for all
individuals at all SNPs}

\item{nc}{The number of computing cores to use. This should never be
more than the number of cores available in your computing environment.
You can determine the maximum number of available cores by running
\code{parallel::detectCores()} in R. This is probably fine for a
personal computer, but some environments are only
able to use fewer. Ask your admins if you are unsure.}

\item{reltol}{The relative tolerance for the stopping criterion.}

\item{pen}{The penalty to be applied to the likelihood. You can think about
this as the prior sample size.}
}
\description{
This function will run \code{\link{ldest}()} iteratively over
all possible pairs of SNPs provided. Support is provided for parallelization
through the doParallel and foreach packages.
}
\examples{
## Simulate genotypes when true correlation is 0
nloci <- 5
nind  <- 100
K <- 6
nc <- 1
genomat <- matrix(sample(0:K, nind * nloci, TRUE), nrow = nloci)

## Estimate LD
rdf <- mldest_geno(genomat = genomat, K = K, nc = nc)

}
\author{
David Gerard
}
