% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldgeno.R
\name{plot.lddf}
\alias{plot.lddf}
\title{Plot the output of \code{\link{mldest_geno}()} or
\code{\link{mldest_genolike}()} using \code{\link[corrplot]{corrplot}()}}
\usage{
\method{plot}{lddf}(
  x,
  element = c("z", "z_se", "D", "D_se", "r2", "r2_se", "r", "r_se", "p_ab", "p_Ab",
    "p_aB", "p_AB"),
  type = c("upper", "full", "lower"),
  diag = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{lddf}, usually created using
either \code{\link{mldest_geno}()} or \code{\link{mldest_genolike}()}.}

\item{element}{Which element of \code{x} should be plot?}

\item{type}{Character, \code{"full"},
\code{"upper"} (default) or \code{"lower"}, display
full matrix, lower triangular or upper
triangular matrix.}

\item{diag}{Logical, whether display the correlation coefficients
on the principal diagonal.}

\item{...}{Additional arguments to pass to
\code{\link[corrplot]{corrplot}()}. See the documentation of that
function for options.}
}
\description{
Uses the \code{\link[corrplot]{corrplot}} R package to visualize
correlation estimates.
}
\author{
David Gerard
}
